pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.1",
    "ipykernel>=6.30.1",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.12",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "wikipedia>=1.4.0",
]
